requirements.txt
pandas
scikit-learn
numpy
pymongo
fo
folium
matplotlib
branca
python-dotenv
python-tk
